+++ js/filling-information/reset-form.js
@@ -1,7 +1,6 @@
 import {markerGroup, createMarker} from '../utils/create-map.js';
 
-const resetForm = (evt) => {
-  evt.preventDefault();
+const resetForm = () => {
   document.querySelector('.ad-form').reset();
   document.querySelector('.map__filters').reset();
   markerGroup.clearLayers();
